+++ src/features/posts/PostService.ts
@@ -6,6 +6,18 @@ import { PostRepository } from "../../models/Post/PostRepository.ts";
 export class PostService {
     constructor(private readonly postRepository: PostRepository) { }
 
+    private ensurePostOwner(post: IPost, userId: string) {
+        const postAuthorId = String(post.author);
+        const loggedUserId = String(userId);
+        if (postAuthorId !== loggedUserId) {
+            throw throwlhos.err_forbidden("You can only change your own posts", {
+                postId: post._id,
+                authorId: post.author,
+                userId,
+            });
+        }
+    }
+
     async create(newPost: IPost, authorId: string) {
         return await this.postRepository.create({
             ...newPost,
@@ -36,19 +48,34 @@ export class PostService {
         return post;
     }
 
-    async updateById(id: string, data: Partial<IPost>) {
-        const post = await this.postRepository.updateById(id, data);
+
+    async updateById(id: string, data: Partial<IPost>, userId: string) {
+        const currentPost = await this.postRepository.findById(id);
+        if (!currentPost) {
+            throw throwlhos.err_notFound("Post not found", { postId: id });
+        }
+        this.ensurePostOwner(currentPost, userId);
+
+        const { title, content, published, tags } = data;
+        const post = await this.postRepository.updateById(id, {
+            title,
+            content,
+            published,
+            tags,
+        });
         if (!post) {
             throw throwlhos.err_notFound("Post not found", { postId: id });
         }
         return post;
     }
 
-    async delete(id: string) {
+    async delete(id: string, userId: string) {
         const post = await this.postRepository.findById(id);
         if (!post) {
             throw throwlhos.err_notFound("Post not found", { postId: id });
         }
+        this.ensurePostOwner(post, userId);
+
         await this.postRepository.delete(id);
     }
 }
